In this lesson 1-2, I have impleted these following : React Dom, Component Definition, Component instantiation and React components.
