+++ src/AddTodoForm.jsx
@@ -1,0 +1,12 @@
+function AddTodoForm(){
+
+    return(
+        <form>
+            <label htmlFor="Title">Title :</label>
+            <input id="todoTitle" type="text"/>
+            <button type="submit">Add</button>
+        </form>
+    );
+}
+
+export default AddTodoForm;
